extract hand sorting from setup move handout and make it a single move in phase bidding
