+++ src/boardgame/phases/bidding.ts
@@ -1,9 +1,12 @@
 /* eslint-disable no-param-reassign */
 import { Ctx, PhaseConfig } from "boardgame.io";
 import { INVALID_MOVE } from "boardgame.io/core";
+import groupBy from "lodash/groupBy";
+import flatten from "lodash/flatten";
 import { WizardState, isSetRound } from "../WizardState";
 import { isValidBid } from "../util/bid";
 import { Phase } from "./phase";
+import { Rank } from "../entities/cards";
 
 export function bid(
   { round, numCards, currentPlayer }: WizardState,
@@ -21,6 +24,32 @@ export function bid(
   ctx.events!.endTurn!();
 }
 
+export function sortCards({ currentPlayer, round }: WizardState): void {
+  if (!round) return;
+
+  const hand = round.hands[currentPlayer];
+  // sort cards on each players hand
+  const groupedHand = groupBy(hand, (card) => {
+    if (card.rank === Rank.Z || card.rank === Rank.N) return card.rank;
+    return card.suit;
+  });
+
+  const sortedGroups = Object.entries(groupedHand)
+    .sort(([keyA, cardsA], [keyB, cardsB]) => {
+      // always sort N left and Z right
+      if (keyA === Rank.N.toString() || keyB === Rank.Z.toString()) return -1;
+      if (keyA === Rank.Z.toString() || keyB === Rank.Z.toString()) return 1;
+      // always sort trump beside Z
+      if (keyA === round.trump?.suit) return 1;
+      if (keyB === round.trump?.suit) return -1;
+      // sort by number of cards of color
+      return cardsA.length - cardsB.length;
+    })
+    .map(([, cardGroup]) => cardGroup.sort((a, b) => a.rank - b.rank));
+  // set sorted hand to state
+  round.hands[currentPlayer] = flatten(sortedGroups);
+}
+
 function endIf({ round }: WizardState): boolean {
   if (!isSetRound(round)) {
     throw new Error("round is not set");
@@ -31,6 +60,7 @@ function endIf({ round }: WizardState): boolean {
 export const bidding: PhaseConfig = {
   moves: {
     bid,
+    sortCards,
   },
   endIf,
   next: Phase.Playing,
